Fix duplicate examine event subscription
